refactor(ai-builder): Repoint eval harness importers to the runner domain modules (no-changelog)
Repoint every harness/runner import (run/ phases, CLI entrypoints, package
barrel, tests incl. vi.mock paths) to the domain modules, delete the
re-export facade, and add the harness module map to ARCHITECTURE.md.
