+++ packages/@n8n/instance-ai/evaluations/__tests__/runner-thread-id.test.ts
@@ -1,8 +1,8 @@
 import { vi } from 'vitest';
 
 import type { N8nClient } from '../clients/n8n-client';
+import { buildWorkflow } from '../harness/build-workflow';
 import type { EvalLogger } from '../harness/logger';
-import { buildWorkflow } from '../harness/runner';
 
 const silentLogger: EvalLogger = {
 	info: () => {},
