+++ packages/@n8n/instance-ai/evaluations/__tests__/verification-artifact.test.ts
@@ -1,7 +1,7 @@
 import type { InstanceAiEvalExecutionResult, InstanceAiEvalNodeResult } from '@n8n/api-types';
 
 import type { WorkflowResponse } from '../clients/n8n-client';
-import { buildVerificationArtifact, selectScenarioWorkflowId } from '../harness/runner';
+import { buildVerificationArtifact, selectScenarioWorkflowId } from '../harness/scenario-execution';
 import type { ExecutionScenario } from '../types';
 
 function makeNodeResult(
