+++ packages/@n8n/instance-ai/evaluations/__tests__/workflow-expected.test.ts
@@ -1,4 +1,4 @@
-import { workflowExpectedForCase } from '../harness/runner';
+import { workflowExpectedForCase } from '../harness/build-workflow';
 
 describe('workflowExpectedForCase', () => {
 	const scenario = {
