+++ packages/@n8n/instance-ai/evaluations/__tests__/write-partial-results.test.ts
@@ -3,8 +3,8 @@ import { existsSync, mkdtempSync, readFileSync, rmSync } from 'node:fs';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 
+import { abortedWorkflowTestCaseResult } from '../harness/cleanup';
 import type { EvalLogger } from '../harness/logger';
-import { abortedWorkflowTestCaseResult } from '../harness/runner';
 import { runEvalAndPersist } from '../run/persist';
 import type { ExecutionScenario, WorkflowTestCase, WorkflowTestCaseResult } from '../types';
