+++ packages/@n8n/instance-ai/evaluations/cli/args.ts
@@ -127,7 +127,7 @@ export interface CliArgs {
 // ---------------------------------------------------------------------------
 
 const cliArgsSchema = z.object({
-	// Keep in sync with DEFAULT_TIMEOUT_MS in harness/runner.ts (and its
+	// Keep in sync with DEFAULT_TIMEOUT_MS in harness/build-workflow.ts (and its
 	// rationale for why this default must stay conservative).
 	timeoutMs: z.number().int().positive().default(900_000),
 	baseUrls: z.array(z.string().url()).min(1).default(['http://localhost:5678']),
