+++ packages/@n8n/instance-ai/evaluations/cli/build-mcp-manifest.ts
@@ -14,9 +14,9 @@ import {
 	stageMcpConfigFromClaudeJson,
 	uniqueProjectScopes,
 } from './mcp-builder';
+import { runWithConcurrency } from '../harness/cleanup';
 import { createLogger } from '../harness/logger';
 import { prebuiltManifestSchema, type PrebuiltManifest } from '../harness/prebuilt-workflows';
-import { runWithConcurrency } from '../harness/runner';
 import { ConversationTurnSchema, DEFAULT_DATASETS } from '../harness/schema';
 import { loadTestCasesFromLangTracer } from '../langtracer/provider';
