+++ packages/@n8n/instance-ai/evaluations/run/direct-driver.ts
@@ -16,7 +16,7 @@ import { type RowSink } from './persist';
 import { reshapeLangSmithRuns, type ReshapeRunRow } from './reshape';
 import { BUILD_ONLY_SCENARIO_NAME, roundRobinCaseRows } from './rows';
 import type { WorkflowTestCaseWithFile } from '../data/workflows';
-import { runWithConcurrency } from '../harness/runner';
+import { runWithConcurrency } from '../harness/cleanup';
 import type { MultiRunEvaluation, WorkflowTestCase, WorkflowTestCaseResult } from '../types';
 
 export interface DirectRunConfig extends Omit<EvalSessionConfig, 'wrap'> {
